+++ boundingbox_test.go
@@ -8,24 +8,9 @@ import (
 	"github.com/EricNeid/go-getosm/internal/verify"
 )
 
-func TestParseTileMode(t *testing.T) {
-	// action
-	res, err := ParseTileMode("vertical")
-	// verify
-	verify.Ok(t, err)
-	verify.Equals(t, TileVertical, res)
-}
-
-func TestParseTileMode_invalidMode(t *testing.T) {
-	// action
-	_, err := ParseTileMode("invalid")
-	// verify
-	verify.NotNil(t, err, "")
-}
-
 func TestReadBoundingBox_tileVertical(t *testing.T) {
 	// action
-	bbs, err := ReadBoundingBox("10,50,11,51", 2, TileVertical)
+	bbs, err := ReadBoundingBox("10,50,11,51", 2, 1)
 	// verify
 	verify.Ok(t, err)
 	verify.Equals(t, 2, len(bbs))
@@ -43,7 +28,7 @@ func TestReadBoundingBox_tileVertical(t *testing.T) {
 
 func TestReadBoundingBox_tileHorizontal(t *testing.T) {
 	// action
-	bbs, err := ReadBoundingBox("10,50,11,51.0", 2, TileHorizontal)
+	bbs, err := ReadBoundingBox("10,50,11,51.0", 1, 2)
 	// verify
 	verify.Ok(t, err)
 	verify.Equals(t, 2, len(bbs))
@@ -57,4 +42,32 @@ func TestReadBoundingBox_tileHorizontal(t *testing.T) {
 	verify.AlmostEquals(t, 11.0, bbs[1].East)
 	verify.AlmostEquals(t, 50.5, bbs[1].South)
 	verify.AlmostEquals(t, 51.0, bbs[1].North)
+}
+
+func TestReadBoundingBox_tileGrid(t *testing.T) {
+	// action
+	bbs, err := ReadBoundingBox("10,50,11,51.0", 2, 2)
+	// verify
+	verify.Ok(t, err)
+	verify.Equals(t, 4, len(bbs))
+
+	verify.AlmostEquals(t, 10.0, bbs[0].West)
+	verify.AlmostEquals(t, 10.5, bbs[0].East)
+	verify.AlmostEquals(t, 50.0, bbs[0].South)
+	verify.AlmostEquals(t, 50.5, bbs[0].North)
+
+	verify.AlmostEquals(t, 10.5, bbs[1].West)
+	verify.AlmostEquals(t, 11.0, bbs[1].East)
+	verify.AlmostEquals(t, 50.0, bbs[1].South)
+	verify.AlmostEquals(t, 50.5, bbs[1].North)
+
+	verify.AlmostEquals(t, 10.0, bbs[2].West)
+	verify.AlmostEquals(t, 10.5, bbs[2].East)
+	verify.AlmostEquals(t, 50.5, bbs[2].South)
+	verify.AlmostEquals(t, 51.0, bbs[2].North)
+
+	verify.AlmostEquals(t, 10.5, bbs[3].West)
+	verify.AlmostEquals(t, 11.0, bbs[3].East)
+	verify.AlmostEquals(t, 50.5, bbs[3].South)
+	verify.AlmostEquals(t, 51.0, bbs[3].North)
 }