+++ log.go
@@ -6,8 +6,10 @@ import (
 	"github.com/op/go-logging"
 )
 
+// Log is the logger for this application.
 var Log = logging.MustGetLogger("gogetosm")
 
+// SetLogLevel sets the log level for the application logger.
 func SetLogLevel(level logging.Level) {
 	logging.SetLevel(level, "gogetosm")
 }